Create a wrapper biblatex to allow BibTeX or Biber support as required
This just does enough to load the real code, which will either be the
'frozen' BibTeX suppport (v1) or the current Biber support (v2).

Any items added to biblatex v2 which appear in any common file
(e.g. biblatex.def) have to be set up here for v1 if necessary, or
of course added to the v1 code.