Update spanish.lbx
Fixed a lot of missing items. There are some gender issues with some entries (whose abbreviations in general are not affected). In those cases, I opted for a «neutral» form, i.e. the noun: for instance, «edición» instead of «editor» or «editora», etc.. If those changes are approved, then we can delete the explanations.