Add a stub blx-bibtex.def
The best plan for BibTeX support code is to put it in a separate
file and to make minimal adjustment to biblatex.sty to read it at
the 'right' point(s). The exact detail for that is still to be decided:
possibly it will need an option, possibly it can be picked up from
the .bbl file. (The twin issues are writing to the .aux/special
.bib file and the areas where cross-compatibility even in a fallback
sense may not work.)

The idea is to add only enough code here to allow BibTeX to work, with
adjustments minimal in biblatex.sty and matching changes in the .bst
file (which is currently as-written by PL).