Writing of control .bib file
Mainly a copy-paste but with "@bibtex" removed in various places.

The sorting key will need some work as the form is currently different
between Biber and BibTeX.

The data model here is hard-coded in the toggles: that should be fine
as this will need to be forced in any case (probably via a reset-then-
deactivate approach).