Enable backend option and input of new .def file
At present this is a simple yes/no: still need to think about
the requirement to distinguish bibtex/bibtex8. Almost certainly don't
want bibtexu to be supported: as it doesn't work properly anyway that
seems silly.

The load point here is aimed at allowing \RequireBiber, etc., to
appear in styles and do the right thing, and also any other degrading
code that gets added to apply to styles. This may yet have to be
somewhat adjusted!