Update .bst file to produce new name format
We now can produce the 'flexible' model output to match Biber.
At the moment, there is no coverage for "terseinits" but this will
be added soon: the format produced by Biber can be obtained (at least
in a fallback sense) using search-replace in BibTeX.

This is almost working with BibTeX as the backend for the simple case:
still needs some coverage for sorting (a no-op with BibTeX but must
be present).