Use \protected@edef in \forcezeros
Get rid of \IfInteger in favour of \protected@edef to be safer for unusual
input cf. https://github.com/plk/biber/issues/237 and
https://github.com/plk/biblatex/issues/768.

To be completely safe exact control of expansion with \expandafter would
be necessary, but that is hardly pretty. Since the involved macros
are only ever used in two specific situations it could be pulled off,
though...