Add (min|max)sortnames and namestrunchandling for BibTeX (#755, #861)
Write the auxiliary -blx.bib file with .bcf version and not .bbl
version. That probably does not matter a lot, but since the -blx.bib
file has a role similar to the .bcf file and implements a subset
of its features, that version seemed more relevant that that of the
.bbl. To get things absolutely right we would need a new version ID,
because the .bcf is more flexible than the -blx.bib. We shall see...