Contributed Serbian Localization
As announced in #937, here's a Serbian (Cyrillic and Latin) localization
for Biblatex.

The files are generated from a common template and YAML string
collection by a script.
All of these files can be found at
<https://gitlab.com/andrejr/biblatex-serbian>.

Please don't modify these files directly as they are generated --
contribute your changes upstream and then send a pull request with the
generated file to this repo.

This is done so we could generate both `serbian.lbx` (Latin) and
`serbianc.lbx` (Cyrillic) from the same set of localization strings and
with the same date format, ordinal format, and other extras.