Implement \AtEveryEntrykey (#934)
This is a hook that is executed whenever an entry key is processed for
a citation (including an explicit \nocite).

We want to pass the entrykey as argument, so we need a hook with
a parameter. Hence we use \apptocmd and not \appto.
Unfortunately, \apptocmd is a bit more complex and has
success and failure code.