Print punctuation with \setunit in numeric-comp (#988)
That makes it easier for the punctuation buffer/tracker.

This commit does  not resolve the underlying problem in #988 that \cite
in bibliographies does not initialise the buffer, but provides
some relief for the problem at hand and is arguably nicer anyway.
Since we are reworking numeric-comp anyway (#900, #963) we might
as well do this.