Allow spaces between args of \DeclareFieldFormat (#586)
Make one more argument of \blx@defformat@i explicit to avoid
spaces hanging around.

In theory this makes things a bit less general, but we can probably live
with that. Currently, the macro is only ever used with one arg, so
this is safe (and I don't expect that this will change).