Load all .lbx files \AtBeginDocument (#1041)
Languages that are loaded on the fly and are not reported via the
official channels by babel/polyglossia are recorded in the .aux file
and loaded on a subsequent run. A rerun warning is emitted if required.

Allows us to roll back e7e7c38a8f9aa08c374275ee1263763fa092680e.